requirements.txt

cloudinary
distro
Django
django-ckeditor
django-compat
django-js-asset
docutils
Flask
gunicorn
lazy-object-proxy
MarkupSafe
mccabe
packaging
parso
pkgconfig
platformdirs
prompt-toolkit
Pygments
pyjnius
pylint
pyparsing
PySDL2
python-dateutil
pytz
requests
six
soupsieve
sqlparse
toml
tzdata

whitenoise

